fix: removing landing page (#334)
