--- src/components/Home/header/Header.svelte
@@ -1,23 +1,0 @@
-<script>
-  import Logo from '$assets/svg/logos/homepage/contentcredentials.svg';
-</script>
-
-<div class="flex w-full justify-center">
-  <div class="w-full max-w-[1440px] px-5 py-4 sm:px-8 lg:py-7 xl:px-[2rem]">
-    <div class="flex flex-row items-center justify-between">
-      <div class="pr-6">
-        <a href="/">
-          <img
-            src={Logo}
-            width="204px"
-            height="28px"
-            alt="content credentials"
-            class=" lg:h-[46px] lg:w-[261px]" />
-        </a>
-      </div>
-      <div>
-        <a href="/verify" class="underline"> Inspect </a>
-      </div>
-    </div>
-  </div>
-</div>
