--- src/components/Home/homepage/Home1.svelte
@@ -1,38 +1,0 @@
-<script>
-  import KeyPoints from './KeyPoints.svelte';
-
-  function scrollToHowItWorks() {
-    const element = document.getElementById('how-it-works');
-
-    if (element) {
-      element.scrollIntoView({ behavior: 'smooth' });
-    }
-  }
-</script>
-
-<div
-  class="flex w-screen max-w-[90rem] flex-col items-center justify-evenly gap-6 px-[20px] pt-[1.5rem] sm:px-[1.5rem]">
-  <div
-    class="flex flex-col items-center justify-center pb-10 pt-20 lg:w-2/3 lg:max-w-[56rem] xl:w-[802px] 2xl:w-[910px]">
-    <h2
-      class="text-center text-home-secondary-title sm:text-home-title-md lg:text-home-title-lg xl:text-hero-title 2xl:text-home-title-2xl">
-      Content Credentials reveal helpful information that's usually hidden
-    </h2>
-    <div class="flex flex-col items-center justify-center pt-10">
-      <p
-        class="text-center text-hero-sub text-gray-900/60 lg:text-hero-sub-tablet 2xl:text-hero-sub-desktop">
-        Critical information about the content you see online is often
-        inaccessible or inaccurate. Content Credentials are a new open
-        technology for revealing answers to your questions about content with a
-        simple click: How was it made? Is it AI-generated? When was it created
-        or edited?
-      </p>
-      <button
-        on:click={scrollToHowItWorks}
-        class="mt-10 flex h-[2.75rem] items-center scroll-smooth rounded-full bg-brand-red px-[1.25rem] py-3 hover:cursor-pointer hover:brightness-[0.9] focus:border-[1px] focus:border-solid focus:border-gray-900">
-        See how it works
-      </button>
-    </div>
-  </div>
-  <KeyPoints />
-</div>
