--- src/components/Home/homepage/Home2.svelte
@@ -1,31 +1,0 @@
-<script>
-  import Image from '$assets/jpg/home1.jpg';
-  import L2Overlay from '../L2Overlay/L2Overlay.svelte';
-</script>
-
-<div class="w-full md:grid md:grid-cols-12 md:gap-4 xl:max-w-[1440px]">
-  <div class="w-full md:col-span-6">
-    <div class="aspect-[3/2]">
-      <L2Overlay
-        imageUrl={Image}
-        class="h-full w-full"
-        alt="Close-up of a colorful butterfly sitting on a sidewalk, with people walking past blurred in the background." />
-    </div>
-  </div>
-  <div
-    class="pt-6 md:col-start-7 md:col-end-13 md:pt-0 lg:col-start-8 lg:col-end-12">
-    <div class="max-w-[676px] md:mx-0">
-      <h3 class="pb-6 text-left text-key-title 2xl:text-key-title-xl">
-        Building trust in what you see online
-      </h3>
-      <p
-        class="text-key-sub-text text-brand-gray lg:text-key-sub-text-lg 2xl:text-key-sub-text-xl">
-        Content Credentials make the origin and history of content available for
-        everyone to access, anytime. With this information at your fingertips,
-        you have the ability to decide if you trust the content you
-        see—understanding what it is and how much editing or manipulation it
-        went through.
-      </p>
-    </div>
-  </div>
-</div>
