--- src/components/Home/homepage/HowItWorks.svelte
@@ -1,244 +1,0 @@
-<script>
-  import Video from '$assets/video/hiw.mp4';
-  import Hiw1 from '$assets/video/how-it-works-1.mp4';
-  import Hiw2 from '$assets/video/how-it-works-2.mp4';
-</script>
-
-<div
-  class="flex h-full w-screen flex-col items-center bg-[#f5f5f5] px-[20px] pb-40 pt-[4rem] sm:px-[1.5rem] md:pt-[6rem] lg:pt-[8rem] xl:pt-[10rem]"
-  id="how-it-works">
-  <h2
-    class="mb-[64px] text-center text-home-secondary-title sm:mb-[80px] sm:text-home-title-md lg:text-home-title-lg xl:text-hero-title 2xl:text-home-title-2xl">
-    How it works
-  </h2>
-  <div class="flex w-full flex-col items-center 2xl:hidden">
-    <div class="">
-      <div
-        class="max-w-[90rem] lg:max-w-[1145px] xl:max-w-[116rem] xl:px-[10rem] xl:pt-[128px] 2xl:pt-[160px]">
-        <div
-          class="flex max-w-[86rem] flex-col items-start justify-between md:flex-row md:pb-[4rem] xl:pb-[4rem]">
-          <div
-            class="flex flex-col items-center gap-4 md:w-1/2 md:flex-row md:items-start md:justify-start">
-            <div class="relative md:mt-[-4px]">
-              <div
-                class="mr-0.5 h-[40px] w-[40px] rounded-full bg-brand-yellow" />
-              <div
-                class="absolute left-[15px] top-[6px] text-hiw-label xl:text-hiw-label-xl">
-                1
-              </div>
-            </div>
-            <h3 class="text-hiw-label md:pr-10 xl:text-hiw-label-xl">
-              Anytime content is captured, created, or edited, Content
-              Credentials can be attached
-            </h3>
-          </div>
-          <p
-            class="mb-[32px] pt-[16px] text-key-sub-text text-brand-gray md:mb-0 md:w-1/2 md:pt-0 lg:pt-0 lg:text-key-sub-text-lg 2xl:text-key-sub-text-xl">
-            Creators can choose to attach Content Credentials to their content,
-            which might include things like whether AI was used or not. Voila!
-            This information is added to the edit history of the
-            content—creating a permanent record that can be confirmed.
-          </p>
-        </div>
-        <video
-          autoplay
-          loop
-          muted
-          playsinline
-          src={Hiw1}
-          class="dropshadow h-full max-h-full w-full max-w-full rounded-xl object-cover lg:rounded-[20px] xl:h-full"
-          alt="A phone in someones hand, displaying a photo of the surrounding snowy landscape and a penguin. The option to enable Content Credentials for the photo is available" />
-      </div>
-      <div
-        class="max-w-[90rem] pt-[96px] lg:max-w-[1145px] xl:max-w-[116rem] xl:px-[10rem] xl:pt-[128px] 2xl:pt-[160px]">
-        <div
-          class="flex max-w-[86rem] flex-col items-start justify-between md:flex-row md:pb-[4rem] xl:pb-[4rem]">
-          <div
-            class="flex w-full flex-col items-center gap-4 md:w-1/2 md:flex-row md:items-start md:justify-start">
-            <div class="relative md:mt-[-4px]">
-              <div
-                class="mr-0.5 h-[40px] w-[40px] rounded-full bg-brand-orange" />
-              <div
-                class="absolute left-[13px] top-[6px] text-hiw-label xl:text-hiw-label-xl">
-                2
-              </div>
-            </div>
-            <h3
-              class="w-full text-left text-hiw-label md:pr-10 xl:text-hiw-label-xl">
-              Content Credentials are viewable across the internet
-            </h3>
-          </div>
-          <p
-            class="mb-[32px] pt-[16px] text-key-sub-text text-brand-gray md:mb-0 md:w-1/2 md:pt-0 lg:pt-0 lg:text-key-sub-text-lg 2xl:text-key-sub-text-xl">
-            Once the content is made available, anyone can view its Content
-            Credentials by clicking the pin, which reveals the most relevant
-            information directly in context.
-          </p>
-        </div>
-        <video
-          autoplay
-          loop
-          muted
-          playsinline
-          src={Hiw2}
-          class="dropshadow h-full max-h-full w-full max-w-full rounded-xl object-cover lg:rounded-[20px] xl:h-full"
-          alt="Penguins in a desert with sand dunes in the background.
-          Content Credentials information is displayed over the
-          image, including that AI was used to generate some of
-          the content, who produced the content, and more" />
-      </div>
-      <!-- Video Component -->
-      <div
-        class="max-w-[90rem] pt-[96px] lg:max-w-[1145px] xl:max-w-[116rem] xl:px-[10rem] xl:pt-[128px] 2xl:pt-[160px]">
-        <div
-          class="flex max-w-[86rem] flex-col items-start justify-between md:flex-row md:pb-[4rem] lg:pb-[3.5rem] xl:pb-[4rem]">
-          <div
-            class="flex w-full flex-col items-center gap-4 md:w-1/2 md:flex-row md:items-start md:justify-start">
-            <div class="relative md:mt-[-4px]">
-              <div class="mr-0.5 h-[40px] w-[40px] rounded-full bg-brand-red" />
-              <div
-                class="absolute left-[13px] top-[5px] text-hiw-label xl:text-hiw-label-xl">
-                3
-              </div>
-            </div>
-            <h3
-              class="w-full text-left text-hiw-label md:pr-10 xl:text-hiw-label-xl">
-              Go deeper to explore the full edit history
-            </h3>
-          </div>
-          <p
-            class="mb-[32px] pt-[16px] text-key-sub-text text-brand-gray md:mb-0 md:w-1/2 md:pt-0 lg:pt-0 lg:text-key-sub-text-lg 2xl:text-key-sub-text-xl">
-            Content Credentials can capture a detailed history of changes over
-            time. The <a href="/verify" class="underline"> Verify </a> feature allows
-            you to explore this information in depth. Check any content to see if
-            it has Content Credentials.
-          </p>
-        </div>
-        <video
-          autoplay
-          loop
-          muted
-          playsinline
-          src={Video}
-          class="dropshadow h-full max-h-full w-full max-w-full rounded-xl object-cover lg:rounded-[20px] xl:h-full"
-          alt="A tree view of the relationship between different images
-          and how they were used together to create a final image
-          over time, available in the Verify feature" />
-      </div>
-    </div>
-  </div>
-
-  <!-- Desktop 1281px+ -->
-  <div class="hidden w-full max-w-[90rem] flex-col items-center 2xl:flex">
-    <div class="grid grid-cols-12">
-      <div class="col-start-2 col-end-6">
-        <div class="flex flex-row justify-start gap-2">
-          <div class="relative md:mt-[-4px]">
-            <div
-              class="mr-0.5 h-[40px] w-[40px] rounded-full bg-brand-yellow xl:h-[44px] xl:w-[44px]" />
-            <div
-              class="absolute left-[16px] top-[5px] text-hiw-label xl:text-hiw-label-xl">
-              1
-            </div>
-          </div>
-          <p class="xl:text-hiw-label-xl">
-            Anytime content is captured, created, or edited, Content Credentials
-            can be attached
-          </p>
-        </div>
-      </div>
-      <div
-        class="col-start-7 col-end-12 text-brand-gray 2xl:text-key-sub-text-xl">
-        Creators can choose to attach Content Credentials to their content,
-        which might include things like whether AI was used or not. Voila! This
-        information is added to the edit history of the content—creating a
-        permanent record that can be confirmed.
-      </div>
-      <div class="col-start-2 col-end-12 mt-[4rem]">
-        <video
-          autoplay
-          loop
-          muted
-          playsinline
-          src={Hiw1}
-          class="dropshadow h-full max-h-full w-full max-w-full rounded-xl object-cover lg:rounded-[20px] xl:h-full"
-          alt="A phone in someones hand, displaying a photo of the
-          surrounding snowy landscape and a penguin. The option
-          to enable Content Credentials for the photo is available" />
-      </div>
-      <div class="col-start-2 col-end-6 mt-[10rem]">
-        <div class="flex flex-row justify-start gap-2">
-          <div class="relative md:mt-[-4px]">
-            <div
-              class="mr-0.5 h-[40px] w-[40px] rounded-full bg-brand-orange xl:h-[44px] xl:w-[44px]" />
-            <div
-              class="absolute left-[15px] top-[5px] text-hiw-label xl:text-hiw-label-xl">
-              2
-            </div>
-          </div>
-          <p class="xl:text-hiw-label-xl">
-            Content Credentials are viewable across the internet
-          </p>
-        </div>
-      </div>
-      <div
-        class="col-start-7 col-end-12 mt-[10rem] text-brand-gray 2xl:text-key-sub-text-xl">
-        Once the content is made available, anyone can view its Content
-        Credentials by clicking the pin, which reveals the most relevant
-        information directly in context.
-      </div>
-      <div class="col-start-2 col-end-12 mt-[4rem]">
-        <video
-          autoplay
-          loop
-          muted
-          playsinline
-          src={Hiw2}
-          class="dropshadow h-full max-h-full w-full max-w-full rounded-xl object-cover lg:rounded-[20px] xl:h-full"
-          alt="A phone in someones hand, displaying a photo of the
-          surrounding snowy landscape and a penguin. The option
-          to enable Content Credentials for the photo is available" />
-      </div>
-      <div class="col-start-2 col-end-6 mt-[10rem]">
-        <div class="flex flex-row justify-start gap-2">
-          <div class="relative md:mt-[-4px]">
-            <div
-              class="mr-0.5 h-[40px] w-[40px] rounded-full bg-brand-red xl:h-[44px] xl:w-[44px]" />
-            <div
-              class="absolute left-[15px] top-[5px] text-hiw-label xl:text-hiw-label-xl">
-              3
-            </div>
-          </div>
-          <p class="xl:text-hiw-label-xl">
-            Go deeper to explore the full edit history
-          </p>
-        </div>
-      </div>
-      <div
-        class="col-start-7 col-end-12 mt-[10rem] text-brand-gray 2xl:text-key-sub-text-xl">
-        Content Credentials can capture a detailed history of changes over time.
-        The <a href="/verify" class="underline"> Verify </a> feature allows you to
-        explore this information in depth, and upload any content to see if it has
-        Content Credentials.
-      </div>
-      <div class="col-start-2 col-end-12 mt-[4rem]">
-        <video
-          autoplay
-          loop
-          muted
-          playsinline
-          src={Video}
-          class="dropshadow h-full max-h-full w-full max-w-full rounded-xl object-cover lg:rounded-[20px] xl:h-full"
-          alt="A tree view of the relationship between different images
-          and how they were used together to create a final image
-          over time, available in the Verify feature" />
-      </div>
-    </div>
-  </div>
-</div>
-
-<style>
-  .dropshadow {
-    filter: drop-shadow(0px 1px 40px #e6e6e6);
-  }
-</style>
