--- src/components/Home/homepage/KeyPoints.svelte
@@ -1,18 +1,0 @@
-<script>
-  import Home2 from './Home2.svelte';
-  import Home3 from './Home3.svelte';
-  import Home4 from './Home4.svelte';
-</script>
-
-<div
-  class="flex max-w-[90rem] flex-col items-center justify-center pb-[4rem] pt-8 md:pb-[6rem] lg:pb-[8rem] xl:pb-[10rem]">
-  <div class="w-full pt-[3rem] md:pt-[5rem] lg:pt-[8rem] xl:pt-[160px]">
-    <Home2 />
-  </div>
-  <div class="w-full pt-[3rem] md:pt-[5rem] lg:pt-[8rem] xl:pt-[160px]">
-    <Home3 />
-  </div>
-  <div class="w-full pt-[3rem] md:pt-[5rem] lg:pt-[8rem] xl:pt-[160px]">
-    <Home4 />
-  </div>
-</div>
