--- src/components/Home/homepage/Movement.svelte
@@ -1,14 +1,0 @@
-<div
-  class="flex flex-col items-center justify-center pb-10 pt-20 sm:px-[1.5rem] lg:w-2/3 lg:max-w-[56rem] xl:w-[802px] 2xl:w-[910px]">
-  <h2
-    class="text-center text-home-secondary-title sm:text-home-title-md xl:text-hero-title 2xl:text-home-title-2xl">
-    Content Credentials are rolling out across the internet
-  </h2>
-  <div class="flex flex-col items-center justify-center pt-10">
-    <p
-      class="text-center text-hero-sub text-gray-900/60 lg:text-hero-sub-tablet 2xl:text-hero-sub-desktop">
-      Expect to see the Content Credentials pin popping up in more and more
-      places as it gets implemented by partners far and wide.
-    </p>
-  </div>
-</div>
