--- src/components/Home/homepage/Powered.svelte
@@ -1,20 +1,0 @@
-<div
-  class="flex flex-col items-center justify-center pb-10 pt-20 sm:px-[1.5rem] lg:w-2/3 lg:max-w-[56rem] xl:w-[802px] 2xl:w-[910px]">
-  <h2
-    class="text-center text-home-secondary-title sm:text-home-title-md xl:text-hero-title 2xl:text-home-title-2xl">
-    Powered by a growing list of organizations
-  </h2>
-  <div class="flex flex-col items-center justify-center pt-10">
-    <p
-      class="text-center text-hero-sub text-gray-900/60 lg:text-hero-sub-tablet 2xl:text-hero-sub-desktop">
-      Content Credentials are being implemented by leading organizations around
-      the world—powered by the C2PA standard and the <a
-        class="underline"
-        href="https://contentauthenticity.org/"
-        target="_blank"
-        rel="nonreferrer">Content Authenticity Initiative</a> , the cross-industry
-      community building open-source tools for Content Credentials to create a more
-      open and transparent internet.
-    </p>
-  </div>
-</div>
