--- src/components/Home/homepage/Trust.svelte
@@ -1,198 +1,0 @@
-<script>
-</script>
-
-<div
-  class="w-full max-w-[90rem] px-[20px] py-[4rem] sm:px-[2rem] md:py-[6rem] xl:py-[10rem]">
-  <div
-    class="flex h-full w-full flex-col items-center justify-center gap-[5rem]">
-    <h2
-      class="text-center text-home-secondary-title sm:text-home-title-md lg:text-home-title-lg xl:text-hero-title 2xl:text-home-title-2xl">
-      What makes Content Credentials trustworthy
-    </h2>
-    <!-- Desktop 1281px+ -->
-    <ul class="hidden w-full max-w-[90rem] grid-cols-12 gap-[1.25rem] 2xl:grid">
-      <li class="col-start-1 col-end-5 h-[27.6875rem] w-full rounded-[20px]">
-        <div class="h-full">
-          <h3
-            class="w-full rounded-t-xl bg-brand-orange px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-            Truly open standards
-          </h3>
-          <div
-            class="h-[365px] rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 lg:rounded-b-[20px]">
-            <p class="2xl:text-key-sub-text-xl">
-              Content Credentials are based on an open technical specification
-              developed and maintained by the <a
-                href="https://c2pa.org/"
-                target="_blank"
-                rel="noopener noreferrer"
-                class="underline">C2PA</a
-              >, a cross-industry standards development organization. You can
-              review the specification anytime.
-            </p>
-          </div>
-        </div>
-      </li>
-      <li class="col-start-5 col-end-9 h-[27.6875rem] w-full rounded-[20px]">
-        <div class="h-full">
-          <h3
-            class="w-full rounded-t-xl bg-brand-red px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-            Clear tampering detection
-          </h3>
-          <div
-            class="text-trust-desc xl:text-trust-desc-xl 2xl:text-trust-desc-2xl h-[365px] rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 lg:rounded-b-[20px]">
-            <p class="2xl:text-key-sub-text-xl">
-              Content Credentials use technology that makes it clear if content
-              has been altered since its Content Credentials were created.
-              Up-to-date context travels with the content, wherever it goes.
-            </p>
-          </div>
-        </div>
-      </li>
-      <li class="col-start-9 col-end-13 h-[27.6875rem] w-full rounded-[20px]">
-        <div class="h-full">
-          <h3
-            class="w-full rounded-t-xl bg-brand-violet px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-            Accountability
-          </h3>
-          <div
-            class="text-trust-desc xl:text-trust-desc-xl 2xl:text-trust-desc-2xl h-[365px] rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 lg:rounded-b-[20px]">
-            <p class="2xl:text-key-sub-text-xl">
-              Devices, software, and organizations that issue Content
-              Credentials identify themselves in the Content Credentials they
-              issue. Issuers therefore stake their reputations on their ability
-              to record information accurately.
-            </p>
-          </div>
-        </div>
-      </li>
-    </ul>
-    <!-- Desktop 1281px+ -->
-
-    <!-- Desktop Triangle Layout 768px - 1280px -->
-    <ul
-      class="hidden w-full max-w-[90rem] grid-cols-12 gap-[1.25rem] md:block 2xl:hidden">
-      <div class="flex flex-col items-center">
-        <div class="grid grid-cols-12 gap-[1.25rem]">
-          <li class="col-start-1 col-end-7 w-full rounded-[20px]">
-            <div class="">
-              <h3
-                class="w-full rounded-t-xl bg-brand-orange px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-                Truly open standards
-              </h3>
-              <div
-                class="text-trust-desc xl:text-trust-desc-xl 2xl:text-trust-desc-2xl h-[248px] rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 lg:rounded-b-[20px]">
-                <p class="text-key-sub-text lg:text-key-sub-text-lg">
-                  Content Credentials are based on an open technical
-                  specification developed and maintained by the <a
-                    href="https://c2pa.org/"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    class="underline">C2PA</a
-                  >, a cross-industry standards development organization. You
-                  can review the specification anytime.
-                </p>
-              </div>
-            </div>
-          </li>
-          <li class="col-start-7 col-end-13 w-full rounded-[20px]">
-            <div class="">
-              <h3
-                class="w-full rounded-t-xl bg-brand-red px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-                Clear tampering detection
-              </h3>
-              <div
-                class="h-[248px] rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 lg:rounded-b-[20px]">
-                <p class="text-key-sub-text lg:text-key-sub-text-lg">
-                  Content Credentials use technology that makes it clear if
-                  content has been altered since its Content Credentials were
-                  created. Up-to-date context travels with the content, wherever
-                  it goes.
-                </p>
-              </div>
-            </div>
-          </li>
-        </div>
-        <div class="grid grid-cols-12 pt-[1.25rem]">
-          <li class="col-start-4 col-end-10 w-full rounded-[20px]">
-            <div class="h-full">
-              <h3
-                class="w-full rounded-t-xl bg-brand-violet px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-                Accountability
-              </h3>
-              <div
-                class="h-[248px] rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 lg:rounded-b-[20px]">
-                <p class="text-key-sub-text lg:text-key-sub-text-lg">
-                  Devices, software, and organizations that issue Content
-                  Credentials identify themselves in the Content Credentials
-                  they issue. Issuers therefore stake their reputations on their
-                  ability to record information accurately.
-                </p>
-              </div>
-            </div>
-          </li>
-        </div>
-      </div>
-    </ul>
-    <!-- Desktop Triangle Layout -->
-
-    <!-- Mobile 768px & under -->
-    <ul
-      class="flex w-full max-w-[90rem] grid-cols-12 flex-col gap-[1.25rem] md:hidden">
-      <li class="col-start-1 col-end-5 w-full rounded-[20px]">
-        <div class="h-full">
-          <h3
-            class="w-full rounded-t-xl bg-brand-orange px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-            Truly open standards
-          </h3>
-          <div
-            class="h-full rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 sm:h-[208px] lg:rounded-b-[20px]">
-            <p class="text-key-sub-text lg:text-key-sub-text-lg">
-              Content Credentials are based on an open technical specification
-              developed and maintained by the <a
-                href="https://c2pa.org/"
-                target="_blank"
-                rel="noopener noreferrer"
-                class="underline">C2PA</a
-              >, a cross-industry standards development organization. You can
-              review the specification anytime.
-            </p>
-          </div>
-        </div>
-      </li>
-      <li class="col-start-5 col-end-9 w-full rounded-[20px]">
-        <div class="h-full">
-          <h3
-            class="w-full rounded-t-xl bg-brand-red px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-            Clear tampering detection
-          </h3>
-          <div
-            class="h-full rounded-b-xl bg-[#f5f5f5] px-[32px] py-6 sm:h-[208px] lg:rounded-b-[20px]">
-            <p class="text-key-sub-text lg:text-key-sub-text-lg">
-              Content Credentials use technology that makes it clear if content
-              has been altered since its Content Credentials were created.
-              Up-to-date context travels with the content, wherever it goes.
-            </p>
-          </div>
-        </div>
-      </li>
-      <li class="col-start-9 col-end-13 w-full rounded-[20px]">
-        <div class="h-full">
-          <h3
-            class="w-full rounded-t-xl bg-brand-violet px-[32px] py-[20px] text-trust-label lg:rounded-t-[20px]">
-            Accountability
-          </h3>
-          <div
-            class="h-full rounded-b-xl bg-[#f5f5f5] px-[32px] sm:h-[208px] lg:rounded-b-[20px]">
-            <p class="text-key-sub-text lg:text-key-sub-text-lg">
-              Devices, software, and organizations that issue Content
-              Credentials identify themselves in the Content Credentials they
-              issue. Issuers therefore stake their reputations on their ability
-              to record information accurately.
-            </p>
-          </div>
-        </div>
-      </li>
-    </ul>
-    <!-- Mobile 991px & under -->
-  </div>
-</div>
